src/steelworks/__init__.py
"""Top-level package for SteelWorks.

We expose commonly used submodules and functions here so that code elsewhere can
import directly from ``steelworks`` instead of reaching into subpackages.
"""

from . import (
    app,
    database,
    data_import,
    lot_utils,
    models,
    repository,
    services,
)

# package versioning (kept in sync with pyproject)
__version__ = "0.1.0"
