pyproject.toml
[tool.poetry]
name = "steelworks"
version = "0.1.0"
description = "Operations reporting tool for manufacturing defects and shipments."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^2.0"
psycopg2-binary = "^2.9"
pandas = "^2.0"
streamlit = "^1.18"
openpyxl = "^3.0"
python-dotenv = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"